stack: fold in #906 and #911, and clear the four 500-LoC breaches they exposed
#906 and #911 BOTH owe a certification campaign — #906 touches
`crates/atlas-plugin/assets/bfcl/requirements.txt` (a measured input: the
scorer's dependencies define the score) and #911 touches 34 perf-path files.
Neither was in this stack, so landing them meant three campaigns of ~8 GPU-hours
instead of one. The campaign already running was an hour in with the expensive
BFCL legs barely started, so aborting and refolding costs ~1 hour and saves ~16.
That is the whole point of certifying a stack rather than a PR.

#905 and #908 are deliberately NOT folded in: zero perf-path files between them,
so they owe no campaign and merge on their own once the metal lane is green.

Four things had to be resolved rather than merged blindly.

1. #911's `e0813d5ffe` ("Tell the client when a swap-IN fails") is SUPERSEDED by
   963e22f684 from this stack's robustness lane. Verified by artifact, not by a
   cherry-pick verdict: `lifecycle.rs:340 restore_swapped_image` already routes
   through `send_error_to_sink` and frees the sequence and unlinks the spill on
   failure — a superset, since it also extracts the single fallible step.
   Skipped, and the two consumers that wanted its richer fixture now use the
   shared one.

2. #911's ci.yml commit carries the SAME metal fix as TheTom's #907 — three
   independent discoveries of one defect (#907, #909, #911). Kept #911's
   explanation because it names the two guards that stop the stub returning
   (`assert-gates-are-wired.py` statically, `metal_backend::tests::
   embedded_kernels` at runtime); the value is identical either way.

3. #906 WAS INCOMPLETE. It adds `soundfile==0.14.0` but never updated
   `provision.rs`'s exact-pin assertion, which is why its own `cargo test
   --workspace` was red. Fixed here, and the assertion stays EXACT rather than
   becoming "contains" — every pin is a scorer input, and one that floats can
   move the score between runs and silently invalidate a stored baseline.

4. FOUR files went over the 500-LoC cap, all pushed there by this stack
   (tests_objects 495->579, preempt_tests 481->510, rate_limiter 457->507,
   validate_tests 441->506). The allow-list requires a rationale AND a tracking
   issue, so they are split, not excused:
     * tests_objects.rs -> + tests_objects_b.rs. Verified byte-exact: 46 tests
       before, 21+25 after, and the recombined text diffs clean against the
       original. A split that silently dropped tests would still have gone green.
     * validate_tests.rs -> + validate_tests_b.rs, same treatment.
     * preempt_tests.rs: my first attempt COPIED its 286-line header into the
       new file, which satisfies the cap by duplicating a fixture — the opposite
       of SSOT. Redone the way the robustness lane recommended: `PreemptStubModel`
       and `active_seq` MOVED into `scheduler/test_support.rs`, whose own header
       already declares it the home for shared scheduler fixtures. 510 -> 245.
     * rate_limiter.rs -> `rate_limiter/identity.rs`. Identity resolution answers
       "who is this", the parent answers "how fast may they go" — a real
       separation rather than a line-count trick. 507 -> 464.

cargo test --workspace: 5618 passed, 0 failed. check, clippy, fmt, prose checker
and certification-selftest all clean.
