+++ .github/record-signers/daa4bc7c19f25eda.pub
@@ -1,0 +1,3 @@
+# Atlas record signer daa4bc7c19f25eda
+# Ed25519 public key, base64. Added automatically on first use.
+KDtajZr8QrCJALzfzymtR+6TonV/9T/+6VsceUtoFi4=
